Feature: Map Save
I created a feature that saves the map level you were on previously and the specific world level. If you saved the game on map level 3 world 2, when you would reload the game and continue with your character you made. You would still be on map level 3 and world level 2